Mon, Dec 26, 2022, 7:47 PM +00:00

+++ week2/pset_2/Readability/readability/readability.c
@@ -23,7 +23,7 @@ int main(void)
     printf("%i sentences\n", sentences);
 
     // Average number of letter per 100 words
-    double avgLetters = letters / words * 100;
+    double avgLetters = double(letters) / double(words) * 100;
     printf("AvgLetters: %f\n", avgLetters);
 
     // Average number of sentenances per 100 words
